Factor KeyboardInput to separate file
